Use chart component in intentios charts
